fix: Action toolbar menu state
