+++ app/client/src/PluginActionEditor/components/PluginActionContextMenu/styles.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+import { MenuContent as ADSMenuContent } from "@appsmith/ads";
+
+export const MenuContent = styled(ADSMenuContent)`
+  z-index: 100;
+`;
